mptcp: Build-Fix: Handle destroying the request-sock correctly
So, now the request-sock is actually in the regular hash-table. So, we
need to destroy it the same way as regular TCP does. (see changes in
mptcp_check_req_master)

In tcp_check_req, only call the MPTCP-functions if own_req is true.
Because, if it is not, that means that we anyways lost the race, so the
socket will be destroyed.

Also, in mptcp_create_master_sk, we now can actually use
inet_ehash_nolisten with the meta_sk as an argument to remove it right
away. This also allows us to get rid of the sock_put. Because,
previously we called inet_unhash, which doesn't do a sock_put when
removing the socket from the hash-table. inet_ehash_nolisten however
handles this correctly. So, we can get rid of the sock_put.

Finally, call sock_rps_save_rxhash in the MPTCP-functions as well.

Fixes: b568f578baca (Merge tag 'v4.4' into mptcp_trunk)
Signed-off-by: Christoph Paasch <cpaasch@apple.com>